requirements.txt
colorama
colorlog
docopt
jsonschema
jmespath
markdown
pyembed
pyembed-markdown
python-dateutil
python-slugify
pytz
pyyaml
